// src/app/models/common/ctopp-composite.ts
export interface  CToppComposite {
      paScore: number;
      paci: string;
      paDesc: string;
      paRank: string;
        
      pmScore: number;
      pmci: string;
      pmDesc: string;
      pmRank: string;
        
      rsnScore: number;
      rsnci: string;
      rsnDesc: string;
      rsnRank: string;
        
      apaScore: number;
      apaci: string;
      apaDesc: string;
      apaRank: string;
  }